+++ macos/Classes/VpnCheckPlugin.swift
@@ -1,20 +1,13 @@
-import Flutter
+import Cocoa
+import FlutterMacOS
 
-extension String: Error {}
-
-
-public class SwiftVpnCheckPlugin: NSObject, FlutterPlugin {
-
+public class VpnCheckPlugin: NSObject, FlutterPlugin {
   public static func register(with registrar: FlutterPluginRegistrar) {
-    let channel = FlutterMethodChannel(name: "vpn_check", binaryMessenger: registrar.messenger())
-    let instance = SwiftVpnCheckPlugin()
+    let channel = FlutterMethodChannel(name: "vpn_check", binaryMessenger: registrar.messenger)
+    let instance = VpnCheckPlugin()
     registrar.addMethodCallDelegate(instance, channel: channel)
   }
 
-    private func notImplemented() throws {
-        throw "Not implemented"
-    }
-    
   public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
     switch call.method {
     case "vpn_active":
@@ -22,7 +15,7 @@ public class SwiftVpnCheckPlugin: NSObject, FlutterPlugin {
         result(isActive)
         break
     default:
-        break
+      result(FlutterMethodNotImplemented)
     }
   }
 }